requirements.txt
streamlit>=1.28.0
google-cloud-aiplatform>=1.38.0
googlemaps>=4.10.0
python-dotenv>=1.0.0
streamlit-js-eval
